# modules/openapi/src/commonMain/kotlin/com/sphereon/oid/fed/openapi/admin-server.yaml
openapi: 3.0.3
info:
  title: OpenID Federation API
  description: This API allows the interaction with federation endpoints to manage and verify entities within a federation.
  contact:
    name: Sphereon
    email: info@sphereon.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0-d40
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/SphereonInt/OpenIDFederationAPI/1.0.0-d40'
  - url: http://localhost:8081
    description: Admin Server
paths:
...
  /logs:
    get:
      summary: Get recent logs
      description: Retrieve the most recent logs from the system
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: limit
          in: query
          description: Maximum number of logs to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Successful retrieval of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Log'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
  /logs/severity/{severity}:
    get:
      summary: Get logs by severity
      description: Retrieve logs filtered by severity level
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: severity
          in: path
          description: Severity level to filter by
          required: true
          schema:
            type: string
            enum: [ Verbose, Debug, Info, Warn, Error, Assert ]
        - name: limit
          in: query
          description: Maximum number of logs to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Successful retrieval of logs by severity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Log'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
  /logs/tag/{tag}:
    get:
      summary: Get logs by tag
      description: Retrieve logs filtered by tag
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: tag
          in: path
          description: Tag to filter by
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of logs to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Successful retrieval of logs by tag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Log'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  '/trust-marks/{trustMarkId}':
    delete:
      summary: Delete a Trust Mark
      description: Delete an existing Trust Mark for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: trustMarkId
          in: path
          required: true
          schema:
            type: string
          description: The identifier of the Trust Mark to be deleted.
      responses:
        '200':
          description: Trust Mark deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Trust Mark deleted successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trust Mark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/received-trust-marks':
    post:
      summary: Create a Received Trust Mark
      description: Create or update a received Trust Mark for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trustMarkTypeId:
                  type: string
                  description: The identifier of the Trust Mark Type.
                  example: https://example.com/trust-mark-type
                jwt:
                  type: string
                  description: The JWT representing the received Trust Mark.
              required:
                - trustMarkTypeId
                - jwt
      responses:
        '201':
          description: Received Trust Mark created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  trustMarkId:
                    type: string
                    description: The unique identifier of the created received Trust Mark.
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List Received Trust Marks
      description: Retrieve all received Trust Marks for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successful fetch of received Trust Marks.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    trustMarkId:
                      type: string
                      description: The unique identifier of the received Trust Mark.
                    trustMarkTypeId:
                      type: string
                      description: The identifier of the Trust Mark Type.
                    issuedJwt:
                      type: string
                      description: The JWT representing the received Trust Mark.
                    receivedAt:
                      type: string
                      format: date-time
                      description: The timestamp when the Trust Mark was received.
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/received-trust-marks/{receivedTrustMarkId}':
    delete:
      summary: Delete a Received Trust Mark
      description: Delete a specific received Trust Mark for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: receivedTrustMarkId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the received Trust Mark to be deleted.
      responses:
        '200':
          description: Received Trust Mark deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Received Trust Mark deleted successfully.
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Received Trust Mark not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    AccountUsername:
      name: X-Account-Username
      in: header
      description: Optional account username to operate on behalf of
      required: false
      schema:
        type: string
  responses:
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: The user does not have permission to perform the requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      x-tags:
        - federation
      required:
        - error
        - message
      properties:
        status:
          type: integer
          x-stoplight:
            id: joax1cgth4uzd
        error:
          type: string
        message:
          type: string
          description: A human-readable short text describing the error.
        timestamp:
          type: integer
          x-stoplight:
            id: qtn6mqbzpjctv
...
    Log:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the log entry
        severity:
          type: string
          enum: [ Verbose, Debug, Info, Warn, Error, Assert ]
          description: The severity level of the log entry
        message:
          type: string
          description: The log message
        tag:
          type: string
          description: The tag/category of the log entry
        timestamp:
          type: integer
          format: int64
          description: The timestamp when the event occurred (epoch milliseconds)
        throwable_message:
          type: string
          description: Error message if an exception occurred
          nullable: true
        throwable_stacktrace:
          type: string
          description: Stack trace if an exception occurred
          nullable: true
        metadata:
          type: object
          description: Additional contextual information
          nullable: true
          additionalProperties: true
      required:
        - id
        - severity
        - message
        - tag
        - timestamp
...